Combined env requirements.txt
alabaster=0.7.12
argh=0.26.2
astroid=2.3.3
atomicwrites=1.3.0
attrs=19.3.0
autopep8=1.5
Babel=2.8.0
backcall=0.1.0
bcrypt=3.1.7
beautifulsoup4=4.8.2
bleach=3.1.3
blis=0.4.1
ca-certificates=2019.11.28
cachetools=4.0.0
catalogue=1.0.0
certifi=2019.11.28
cffi=1.14.0
chardet=3.0.4
checksumdir=1.1.7
click=7.1.1
cloudpickle=1.3.0
colorama=0.4.3
coverage=4.5.4
cryptography=2.8
cymem=2.0.3
cython-blis=0.4.1
Cython=0.29.15
diff-match-patch=20181111
decorator=4.4.2
defusedxml=0.6.0
deprecation=2.0.7
dialogflow=0.7.2
diff-match-patch=20181111
docopt=0.6.2
docutils=0.16
en-core-web-sm=2.2.5
entrypoints=0.3
flake8=3.7.9
Flask=1.1.1
future=0.18.2
google-api-core=1.16.0
googleapis-common-protos=1.51.0
google-auth=1.11.2
google-cloud-core=1.3.0
google-cloud-storage=1.26.0
google-resumable-media=0.5.0
grpcio=1.27.2
html2text=2020.1.16
idna=2.9
imagesize=1.2.0
importlib-metadata=1.5.0
intel-openmp=2019.4
intervaltree=3.0.2
ipykernel=5.1.4
ipython=7.13.0
ipython-genutils=0.2.0
isort=4.3.21
itsdangerous=1.1.0
jedi=0.14.1
Jinja2=2.11.1
joblib=0.14.1
jsonschema=3.2.0
jupyter-client=6.0.0
jupyter-core=4.6.3
keyring=21.1.1
keras-applications=1.0.8
keras-base=2.3.1
keras-gpu=2.3.1
keras-preprocessing=1.1.0
lazy-object-proxy=1.4.3
lz4-c=1.8.3
m2w64-gcc-libgfortran=5.3.0
m2w64-gcc-libs=5.3.0
m2w64-gcc-libs-core=5.3.0
m2w64-gmp=6.1.0
m2w64-libwinpthread-git=5.0.0.4634.697
matplotlib-base=3.2.1
MarkupSafe=1.1.1
mccabe=0.6.1
mkl-service=2.3.0
mistune=0.8.4
mock=2.0.0
more-itertools=8.2.0
msys2-conda-epoch=20160418
murmurhash=1.0.0
nbconvert=5.6.1
nbformat=5.0.4
ndg-httpsclient=0.5.1
nltk=3.4.4
num2words=0.5.10
numpy=1.18.1
numpydoc=0.9.2
packaging=20.1
pandas=1.0.1
pandocfilters=1.4.2
paramiko=2.7.1
parso=0.5.2
pathtools=0.1.2
pbr=5.4.4
pexpect=4.8.0
pickle-mixin=1.0.2
pickleshare=0.7.5
plac=0.9.6
pluggy=0.12.0
preshed=3.0.2
prompt-toolkit=3.0.4
protobuf=3.11.3
psutil=5.7.0
pusher=2.1.4
pyaml=19.12.0
pyasn1=0.4.8
pyasn1-modules=0.2.8
pycodestyle=2.5.0
pycparser=2.2
pydocstyle=5.0.2
pydot=1.4.1
pyflakes=2.1.1
Pygments=2.6.1
pylint=2.4.4
PyNaCl=1.3.0
pyOpenSSL=19.1.0
pyparsing=2.4.6
PyQt5=5.12.3
PyQt5-sip=4.19.18
PyQtWebEngine=5.12.1
pyrsistent=0.15.7
PySocks=1.7.1
python-crfsuite=0.9.6
python-dateutil=2.8.1
python-jsonrpc-server=0.3.4
python-language-server=0.31.9
python-telegram-bot=12.4.2
pywin32-ctypes=0.2.0
PyYAML=5.3.1
pyzmq=19.0.0
QDarkStyle=2.8
QtAwesome=0.7.0
qtconsole=4.7.1
QtPy=1.9.0
requests=2.23.0
requests-oauthlib=1.2.0
rope=0.16.0
rsa=4
scikit-learn=0.22.1
scipy=1.4.1
six=1.14.0
sklearn-crfsuite=0.3.6
snips-nlu=0.20.2
snips-nlu-parsers=0.4.3
snips-nlu-utils=0.9.1
snowballstemmer=2.0.0
sortedcontainers=2.1.0
soupsieve=1.9.4
spacy=2.2.3
Sphinx=2.4.4
sphinxcontrib-applehelp=1.0.2
sphinxcontrib-devhelp=1.0.2
sphinxcontrib-htmlhelp=1.0.3
sphinxcontrib-jsmath=1.0.1
sphinxcontrib-qthelp=1.0.3
sphinxcontrib-serializinghtml=1.1.4
spyder-kernels=1.8.1
typed-ast=1.4.1
tabulate=0.8.6
tensorflow-base=2.1.0
tensorflow-estimator=2.1.0
tensorflow-gpu=2.1.0
spyder-kernels=1.8.1 
testpath=0.4.4
thinc=7.3.0
tornado=6.0.4
tqdm=4.43.0
traitlets=4.3.3
treedict=0.2.2
typed-ast=1.4.1
ujson=1.35
urllib3=1.25.7
wasabi=0.6.0
watchdog=0.10.2
wcwidth=0.1.8
webencodings=0.5.1
Werkzeug=1.0.0
wincertstore=0.2
win-inet-pton=1.1.0
wrapt=1.12.1
yapf=0.29.0
zipp=3.1.0
zope.interface=4.7.1
pytorch
torchvision
cudatoolkit=10.1
